refactor(user-auth):AuthContext 전역 상태에 userProfile 추가
